move native_service_config out of native_scheduler
