+++ paasta_itests/steps/paasta_native_steps.py
@@ -21,8 +21,8 @@ from paasta_tools.frameworks.adhoc_scheduler import AdhocScheduler
 from paasta_tools.frameworks.native_scheduler import create_driver
 from paasta_tools.frameworks.native_scheduler import LIVE_TASK_STATES
 from paasta_tools.frameworks.native_scheduler import NativeScheduler
-from paasta_tools.frameworks.native_scheduler import NativeServiceConfig
 from paasta_tools.frameworks.native_scheduler import TASK_RUNNING
+from paasta_tools.frameworks.native_service_config import NativeServiceConfig
 from paasta_tools.native_mesos_scheduler import main
 from paasta_tools.native_mesos_scheduler import paasta_native_services_running_here
 from paasta_tools.utils import load_system_paasta_config
