+++ tests/frameworks/test_adhoc_scheduler.py
@@ -8,6 +8,7 @@ from mesos.interface import mesos_pb2
 from paasta_tools import utils
 from paasta_tools.frameworks import adhoc_scheduler
 from paasta_tools.frameworks import native_scheduler
+from paasta_tools.frameworks.native_service_config import NativeServiceConfig
 
 
 @pytest.fixture
@@ -51,7 +52,7 @@ class TestAdhocScheduler(object):
         cluster = "cluster"
 
         service_configs = [
-            native_scheduler.NativeServiceConfig(
+            NativeServiceConfig(
                 service=service_name,
                 instance=instance_name,
                 cluster=cluster,
@@ -117,7 +118,7 @@ class TestAdhocScheduler(object):
         cluster = "cluster"
 
         service_configs = [
-            native_scheduler.NativeServiceConfig(
+            NativeServiceConfig(
                 service=service_name,
                 instance=instance_name,
                 cluster=cluster,
